Update zombie conifg and loader.

See merge request compile-or-die/group-18!24

+++ core/src/main/java/com/compileordie/pvz2/models/entities/zombies/variants/vehicle/ArcadeZombie.java
@@ -6,9 +6,8 @@ import com.compileordie.pvz2.models.entities.zombies.types.ZombieType;
 
 public class ArcadeZombie extends VehicleZombie {
     public static final int waveCost = 600;
-    public ArcadeZombie(int health, double speed, int attackPower, int row, double startX,
-                        double x, double y, double xSpeed, double ySpeed, double delta, int bucketHeadHealth) {
-        // ابتدا فرستادن اطلاعات به کلاس والد همراه با ساخت آبجکت آرکید ماشین با مختصات زامبی
+    public ArcadeZombie(double health, double speed, int attackPower, int row, double startX,
+                        double x, double y, double xSpeed, double ySpeed, double delta, double bucketHeadHealth) {
         super(health, speed, attackPower, row, startX, x, y, xSpeed, ySpeed,
             new ArcadeMachine(bucketHeadHealth, x, y), ZombieType.ARCADE_ZOMBIE);
     }
